pyproject.toml
[project]
name = "watermark"
version = "0.1.0"
description = "图片水印添加工具"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Pillow>=10.0.0",
    "argparse"
]

[project.scripts]
watermark = "main:main"
